% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assets.R
\name{copy_assets}
\alias{copy_assets}
\title{Copy JS / CSS resources in local pkgdown website}
\usage{
copy_assets(snippets = TRUE, holder = TRUE, .overwrite = TRUE)
}
\arguments{
\item{snippets}{Copy the snippets.js script (A custom script offering the
\code{loadSnippet()} JS function) and the required \code{ace.js} library in version
1.2.3.}

\item{holder}{Copy \code{holder.js} version 2.9.0 \url{http://holderjs.com/}. Useful
to add placeholder images.}

\item{.overwrite}{Should it overwrite existing copy ? \code{TRUE} by default
assuming git tracking is used.}
}
\description{
This is a helper to retrieve locally some common resources accross pkgdown
website. They will be copied from \code{quillt} package installation to
\code{pkgdown/assets} folder for \href{https://pkgdown.r-lib.org/articles/customise.html#additional-html-and-files}{website customisation}.
}
\section{Example of usage}{
\itemize{
\item See \href{https://rstudio.github.io/flexdashboard/articles/using.html}{flexdashboard website for usage}
}
}

\examples{
\dontrun{
# do not need holder.js
copy_assets(holder = FALSE)

# Don't overwrite
copy_assets(.overwrite = FALSE)
}
}
