% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{examples}
\alias{examples}
\title{Generate HTML for examples}
\usage{
examples(
  yml = "examples.yml",
  caption = TRUE,
  showcaseOnly = FALSE,
  shinyOnly = FALSE
)
}
\description{
Generate HTML for examples
}
