% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thumbnails.R
\name{thumbnail}
\alias{thumbnail}
\title{Generate HTML for a 4-wide bootstrap thumbnail}
\usage{
thumbnail(title, img, href, caption = TRUE, source = NULL)
}
\arguments{
\item{title}{Used as title for the link tooltip and as caption is \code{caption = TRUE}.}

\item{img}{link to the image to use in the thumbnail.}

\item{href}{link to use when the image is clicked.}

\item{caption}{if \code{FALSE}, only the image will be shown without a caption below.}

\item{source}{link to a source repository. if \code{caption = TRUE}, will be added
as a clickable link of the form \code{(Source)}}
}
\description{
Generate HTML for a 4-wide bootstrap thumbnail
}
