% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templates.R
\name{use_quillt_github_labels}
\alias{use_quillt_github_labels}
\title{Manage Github issue labels}
\usage{
use_quillt_github_labels(delete_default = TRUE)
}
\arguments{
\item{delete_default.}{\code{TRUE} (default) will delete default GH labels not in use with any issue.}
}
\description{
This function can create new labels, update colors and description, and optionally delete default Github labels.
\code{use_quillt_github_labels()} is similar with \code{usethis::use_tidy_labels()}
with the following differences:
\itemize{
\item It will not rename \code{questions} to \code{reprex} but keep both.
\item Add a few labels : \verb{RStudio IDE}, \code{pandoc} and \code{knitr} used to label issue regarding upstream tools
}
}
