% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assets.R
\name{copy_pkgdown_templates}
\alias{copy_pkgdown_templates}
\title{Copy pkgdown template in local pkgdown website}
\usage{
copy_pkgdown_templates()
}
\description{
This is a helper to retrieve locally some specific template to tweak pkgdown
website. They will be copied from \code{quillt} package installation to
\code{pkgdown/templates} folder for \href{https://pkgdown.r-lib.org/articles/customise.html#additional-html-and-files}{website customisation}.
}
