% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thumbnails.R
\name{examples}
\alias{examples}
\title{Generate HTML for examples}
\usage{
examples(
  yml = "examples.yml",
  caption = TRUE,
  showcaseOnly = FALSE,
  shinyOnly = FALSE
)
}
\arguments{
\item{yml}{Path to the YAML file.}

\item{caption}{if \code{FALSE}, the title won't be shown as caption below the image.}

\item{showcaseOnly}{if \code{TRUE}, only the element with \code{showcase: TRUE} in YAML will be shown.}

\item{shinyOnly}{if \code{TRUE}, only the element with \code{shiny: TRUE} in YAML will be shown.}
}
\description{
Used typically in \strong{quillt} website, this will read a YAML file containing
the expecting information and build the HTML to be included in the \code{examples.Rmd} vignette.
}
\section{YAML structure}{
The YAML should be following this example\if{html}{\out{<div class="yaml">}}\preformatted{- title: title to use as tooltip and caption # mandatory
  href: link to use when the image or caption is clicked # mandatory
  img: link to the image to use in thumbnail # mandatory
  source: link to the source repo to link to from the caption # optional
  showcase: set to FALSE if you want to filter out # optional
  shiny: set to TRUE if shiny example # optional
}\if{html}{\out{</div>}}

See existing pkgdown website using \strong{quillt} for examples
}

