% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templates.R
\name{setup-helpers}
\alias{setup-helpers}
\alias{use_coc}
\alias{use_contributing}
\alias{use_github_action_quillt_pkgdown}
\alias{use_quillt_pkgdown}
\title{Helpers for project setup with quillt}
\usage{
use_coc()

use_contributing()

use_github_action_quillt_pkgdown()

use_quillt_pkgdown(config_file = "_pkgdown.yml", destdir = "reference")
}
\arguments{
\item{config_file}{Path to the pkgdown yaml config file - could be set to be
\verb{pkgdown/} subfolder.}

\item{destdir}{Target directory for pkgdown docs. By default, it will be in
\code{reference} sub directory for R Markdown related package using \strong{quillt}.}
}
\description{
These helper are mostly inspired by the one in \strong{usethis} from the tidyverse
team and are useful to setup a new pkgdown project with \strong{quillt}
}
\details{
\itemize{
\item \code{use_coc()}: use \code{usethis::use_tidy_coc()} and put _CODE\_OF\\emph{CONDUCT.md} in
\verb{.github/} directory setting the contact email to Rstudio one.
\item \code{use_contributing}: Add a \emph{CONTRIBUTING.md} file in \verb{.github/} following a
template in \strong{quillt}. Inspired by \code{usethis::use_tidy_contributing()}.
\item \code{use_github_action_quillt_pkgdown}: Set a Github Action workflow to build a
pkgdown website and deploy to Netlify. The deployment action use allow a main
deploy branch and PR previews. This action will be triggered on PR only for
branch targetting master and with a name starting with \verb{pkgdown/}
}
}
