% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{thumbnails}
\alias{thumbnails}
\title{Generate HTML for several rows of 4-wide bootstrap thumbnails}
\usage{
thumbnails(thumbs)
}
\description{
Generate HTML for several rows of 4-wide bootstrap thumbnails
}
