% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thumbnails.R
\name{thumbnails}
\alias{thumbnails}
\title{Generate HTML for several rows of 3-wide bootstrap thumbnails}
\usage{
thumbnails(thumbs)
}
\arguments{
\item{thumbs}{A list of thumbnail HTML components, usually built with \code{thumbnail()}}
}
\description{
Generate HTML for several rows of 3-wide bootstrap thumbnails
}
